+++ src/client.ts
@@ -1,5 +1,6 @@
 import { EventEmitter } from 'events';
-import queryString from 'query-string';
+
+import { buildQuerystring, stripTokenFromUrl } from './utils';
 
 export interface OidcJwtClientOptions {
   url: string;
@@ -104,20 +105,6 @@ export interface OidcJwtClient {
    * Remove a listener that was added through addSessionListener().
    */
   removeSessionListener(callback: () => void): void;
-}
-
-function buildQuerystring(params: Record<string, string>): string {
-  return Object.keys(params)
-    .map((k) => encodeURIComponent(k) + '=' + encodeURIComponent(params[k]))
-    .join('&');
-}
-
-function stripTokenFromUrl(href: string): string {
-  //= > {url: 'https://foo.bar', query: {foo: 'bar'}}
-  const parsedUrl = queryString.parseUrl(href, { parseFragmentIdentifier: true });
-  const { url, query, fragmentIdentifier } = parsedUrl;
-  const { token, ...params } = query;
-  return queryString.stringifyUrl({ url, query: params, fragmentIdentifier });
 }
 
 const SessionChangedEvent = 'session_changed';
